doc updates (ranking) + refactoring
