+++ docs/synthetic.md
@@ -2,10 +2,4 @@
 
 Synthetic queries are generated continuously via an LLM. Currently, the LLM generates an article based off of a few Wikipedia pages.
 
-The synthetic synapse generation function can be found here:
-
-https://github.com/VectorChat/chunking_subnet/blob/cd54f5ebc082613bcfbb8326b3f06800a272a6bf/chunking/validator/task_api.py#L435-L455
-
-The logic for generating the synthetic document can be found here:
-
-https://github.com/VectorChat/chunking_subnet/blob/cd54f5ebc082613bcfbb8326b3f06800a272a6bf/chunking/validator/task_api.py#L287-L400
+The functions that generate the synthetic queries can be found at [chunking/utils/synthetic/synthetic.py](https://github.com/VectorChat/chunking_subnet/blob/main/chunking/utils/synthetic/synthetic.py).
